+++ Project_8_Moonshot/Moonshot/Moonshot/ContentView.swift
@@ -10,37 +10,26 @@ import SwiftUI
 struct ContentView: View {
     let astronauts: [String: Astronaut] = Bundle.main.decode("astronauts.json")
     let missions: [Mission] = Bundle.main.decode("missions.json")
-
-    let colums = [
-        GridItem(.adaptive(minimum: 150))
-    ]
+    @State private var gridItemWidth: CGFloat = 150
 
     var body: some View {
         NavigationView {
             ScrollView {
-                LazyVGrid(columns: colums) {
+                LazyVGrid(columns: [GridItem(.adaptive(minimum: gridItemWidth))]) {
                     ForEach(missions) { mission in
                         NavigationLink {
                             MissionView(mission: mission, astronauts: astronauts)
                         } label: {
-                            VStack {
-                                Image(mission.image)
-                                    .resizable()
-                                    .scaledToFit()
-                                    .frame(width: 100, height: 100)
-                                    .padding()
-
-                                VStack {
-                                    Text(mission.displayName)
-                                        .font(.headline)
-                                        .foregroundColor(.white)
-                                    Text(mission.formattedLaunchDate)
-                                        .font(.caption)
-                                        .foregroundColor(.white.opacity(0.5))
+                            Group {
+                                if gridItemWidth == .infinity {
+                                    HStack {
+                                        makeGroup(for: mission)
+                                    }
+                                } else {
+                                    VStack {
+                                        makeGroup(for: mission)
+                                    }
                                 }
-                                .padding(.vertical)
-                                .frame(maxWidth: .infinity)
-                                .background(.lightBackground)
                             }
                             .clipShape(RoundedRectangle(cornerRadius: 10))
                             .overlay(RoundedRectangle(cornerRadius: 10)
@@ -53,8 +42,47 @@ struct ContentView: View {
             .navigationTitle("Moonshot")
             .background(.darkBackground)
             .preferredColorScheme(.dark)
+            .toolbar {
+                ToolbarItem(id: "ChangeGridType", placement: .primaryAction) {
+                    Button {
+                        withAnimation {
+                            gridItemWidth = gridItemWidth == .infinity ? 150 : .infinity
+                        }
+                    } label: {
+                        Image(systemName: gridItemWidth == .infinity ? "square.grid.2x2" : "rectangle.grid.1x2")
+                    }
+
+                }
+            }
         }
         .navigationViewStyle(.stack)
+    }
+
+    private func makeGroup(for mission: Mission) -> some View {
+        Group {
+            Image(mission.image)
+                .resizable()
+                .scaledToFit()
+                .frame(width: 100, height: 100)
+                .padding()
+
+            makeMissionTextsView(for: mission)
+        }
+    }
+
+    private func makeMissionTextsView(for mission: Mission) -> some View {
+        VStack {
+            Text(mission.displayName)
+                .font(.headline)
+                .foregroundColor(.white)
+            Text(mission.formattedLaunchDate)
+                .font(.caption)
+                .foregroundColor(.white.opacity(0.5))
+        }
+        .padding(.vertical)
+        .frame(maxWidth: .infinity)
+        .background(.lightBackground)
+        
     }
 }
