+++ Project_8_Moonshot/Moonshot/Moonshot/MissionCrewView.swift
@@ -1,0 +1,54 @@
+//
+//  MissionCrewView.swift
+//  Moonshot
+//
+//  Created by Darya Viter on 02.06.2022.
+//
+
+import SwiftUI
+
+struct MissionCrewView: View {
+    let crew: [MissionView.CrewMember]
+    
+    var body: some View {
+        ScrollView(.horizontal, showsIndicators: false) {
+            HStack {
+                ForEach(crew, id: \.role) { crewMember in
+                    NavigationLink {
+                        AstronautView(astronaut: crewMember.astronaut)
+                    } label: {
+                        HStack {
+                            Image(crewMember.astronaut.id)
+                                .resizable()
+                                .frame(width: 104, height: 72)
+                                .clipShape(Capsule())
+                                .overlay(
+                                    Capsule()
+                                        .strokeBorder(.white)
+                                )
+
+                            VStack(alignment: .leading) {
+                                Text(crewMember.astronaut.name)
+                                    .foregroundColor(.primary)
+                                    .font(.headline)
+
+                                Text(crewMember.role)
+                                    .foregroundColor(.secondary)
+                            }
+                        }
+                        .padding(.horizontal)
+                    }
+
+                }
+            }
+        }
+    }
+}
+
+struct MissionCrewView_Previews: PreviewProvider {
+    static let astronauts: [String: Astronaut] = Bundle.main.decode("astronauts.json")
+    static let crew: [MissionView.CrewMember] = [ .init(role: "Commander", astronaut: astronauts["armstrong"]!) ]
+    static var previews: some View {
+        MissionCrewView(crew: crew)
+    }
+}
