+++ Project_8_Moonshot/Moonshot/Moonshot/MissionView.swift
@@ -1,0 +1,63 @@
+//
+//  MissionView.swift
+//  Moonshot
+//
+//  Created by Darya Viter on 02.06.2022.
+//
+
+import SwiftUI
+
+struct MissionView: View {
+    struct CrewMember {
+        let role: String
+        let astronaut: Astronaut
+    }
+
+    let mission: Mission
+    let crew: [CrewMember]
+
+    var body: some View {
+        GeometryReader { geometry in
+            ScrollView {
+                VStack {
+                    Image(mission.image)
+                        .resizable()
+                        .scaledToFit()
+                        .frame(maxWidth: geometry.size.width * 0.6)
+
+                    Text(mission.launchDate?.formatted(date: .long, time: .omitted) ??  "N/A")
+                        .padding()
+
+                    MissionDescriptionView(missionDescription: mission.description)
+
+                    MissionCrewView(crew: crew)
+                }
+                .padding(.bottom)
+            }
+        }
+        .navigationTitle(mission.displayName)
+        .navigationBarTitleDisplayMode(.inline)
+        .background(.darkBackground)
+    }
+
+    init(mission: Mission, astronauts: [String: Astronaut]) {
+        self.mission = mission
+        crew = mission.crew.map { member in
+            if let astronaut = astronauts[member.name] {
+                return CrewMember(role: member.role, astronaut: astronaut)
+            } else {
+                fatalError("Missing \(member.name)")
+            }
+        }
+    }
+}
+
+struct MissionView_Previews: PreviewProvider {
+    static let missions: [Mission] = Bundle.main.decode("missions.json")
+    static let astronauts: [String: Astronaut] = Bundle.main.decode("astronauts.json")
+
+    static var previews: some View {
+        MissionView(mission: missions[0], astronauts: astronauts)
+            .preferredColorScheme(.dark)
+    }
+}
